refactor lazy computations to seperate file
